Reorganized React Server App
Instead of bundling the whole node/express server, the react
middle is included in the bundle. It exports middleware to express
which can now be managed separately.
